% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_glm.R
\name{estimate_houbv_params}
\alias{estimate_houbv_params}
\title{Estimate OU Mean-Reversion Parameters}
\usage{
estimate_houbv_params(
  deltaY,
  prob0,
  prob1,
  gamma0_init,
  gamma1_init,
  phi0_init,
  h,
  dt,
  sigma_s
)
}
\arguments{
\item{deltaY}{Observed increments.}

\item{h}{Linear drift.}

\item{dt}{Time increments.}

\item{sigma_s}{Observation noise volatility.}

\item{p}{Filtered state probabilities.}

\item{a}{OU drift parameters.}

\item{phi0}{Initial value of the process.}
}
\value{
An \code{optim} object containing parameter estimates.
}
\description{
Estimates the mean-reversion parameters of the bounded-variation
OU process via likelihood maximization.
}
