% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_glm.R
\name{loglik_phi}
\alias{loglik_phi}
\title{Log-Likelihood for Bounded-Variation OU Model}
\usage{
loglik_phi(deltaY, phi, h, dt, sigma_s)
}
\arguments{
\item{deltaY}{Observed increments.}

\item{phi}{Estimated bounded-variation process.}

\item{h}{Linear drift.}

\item{dt}{Time increments.}

\item{sigma_s}{Observation noise volatility.}
}
\value{
Numeric value of the negative log-likelihood.
}
\description{
Computes the negative log-likelihood of the observation process
given the bounded-variation OU dynamics.
}
