% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{step1_cthmm}
\alias{step1_cthmm}
\title{Estimate Continuous-Time Hidden Markov Model}
\usage{
step1_cthmm(deltaY, time, trans_init, dt, hidden_model0, hidden_model1)
}
\arguments{
\item{deltaY}{Numeric vector of observed increments.}

\item{time}{Observation times.}

\item{trans_init}{Initial generator matrix.}

\item{dt}{Time increment.}

\item{hidden_model0}{Hidden model estimated in step 1 to get prior probabilities for the first state. Should be of form hmmNorm(mean,sd).}

\item{hidden_model1}{Hidden model estimated in step 1 to get prior probabilities for the second state. Should be of form hmmNorm(mean,sd).}

\item{obs_noise}{Observation noise.}
}
\value{
A list containing the fitted HMM, estimated generator matrix,
and filtered state probabilities.
}
\description{
Estimates a two-state continuous-time hidden Markov model
with Gaussian emissions.
}
\details{
This function relies on the \pkg{msm} package to compute
likelihoods for continuous-time Markov chains using matrix exponentials.
}
\seealso{
\code{\link[msm]{msm}}
}
