% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_glm.R
\name{estimate_houbv_params}
\alias{estimate_houbv_params}
\title{Estimate OU Parameters}
\usage{
estimate_houbv_params(
  deltaY,
  prob0,
  prob1,
  gamma0_init,
  gamma1_init,
  phi0_init,
  h,
  dt,
  sigma_s
)
}
\arguments{
\item{deltaY}{Observed increments.}

\item{prob0}{Mixture state 1 probabilities.}

\item{prob1}{Mixture state 2 probabilities.}

\item{gamma0_init}{Initial value for gamma0.}

\item{gamma1_init}{Initial value for gamma1.}

\item{phi0_init}{Initial value of the process.}

\item{h}{Linear drift.}

\item{dt}{Time increment.}

\item{sigma_s}{Observation noise volatility.}
}
\value{
An \code{optim} object containing parameter estimates.
}
\description{
Estimates the parameters of the bounded-variation
OU process via likelihood maximization.
}
