% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{fit_cthoubv}
\alias{fit_cthoubv}
\title{Fit a Continuous-Time Hidden Ornstein-Uhlenbeck process of bounded variation}
\usage{
fit_cthoubv(
  deltaY,
  time,
  h,
  trans_init,
  gamma0_init,
  gamma1_init,
  phi0_init,
  sigma_s,
  dt,
  hidden_model0,
  hidden_model1,
  mode = "mixture_probs",
  max_iter = 50,
  tol = 1e-04
)
}
\arguments{
\item{deltaY}{Numeric vector of observed increments \eqn{\Delta Y_{t_n}}.}

\item{time}{Numeric vector of observation times \eqn{t_n}.}

\item{h}{Linear drift.}

\item{trans_init}{Initial generator matrix for CTMC.}

\item{gamma0_init}{Initial value for gamma0.}

\item{gamma1_init}{Initial value for gamma1.}

\item{phi0_init}{Initial value of the process.}

\item{sigma_s}{Observation noise volatility.}

\item{dt}{Time increment.}

\item{hidden_model0}{Hidden model estimated in step 1 to get prior probabilities for the first state. Should be of form hmmNorm(mean,sd).}

\item{hidden_model1}{Hidden model estimated in step 1 to get prior probabilities for the second state. Should be of form hmmNorm(mean,sd).}

\item{mode}{Either "mixture_probs" for "MAP_probs", meaning that for each time point we consider the most likely state.}

\item{max_iter}{Maximum number of EM iterations.}

\item{tol}{Convergence tolerance for the negative log-likelihood.}
}
\value{
A list containing:
\describe{
\item{gamma0_estim}{Estimated parameter gamma0.}
\item{gamma1_estim}{Estimated parameter gamma1.}
\item{phi0_estim}{Estimated initial state.}
\item{phi}{Estimated hidden process.}
\item{MC_res}{Estimated Markov chain.}
}
}
\description{
Fits a continuous-time hidden OUbv model
using an EM-style iterative procedure.
The latent Markov chain is estimated in continuous time, and
the bounded-variation process is inferred using filtered
state probabilities.
}
\details{
The algorithm alternates between:
\enumerate{
\item Estimating the continuous-time hidden Markov chain using
a Gaussian emission model.
\item Updating the bounded-variation process and optimizing
OU parameters via likelihood maximization.
}
}
\references{
Baum, L. E., et al. (1970). A maximization technique occurring in
the statistical analysis of probabilistic functions of Markov chains.
}
